Connect: enable not passing address
If an application is sure it's `attach()`` will succeed, it doesn't
have to supply `addr` straight away. It may rather postpone getting
address until it's actually needed: when application UUID changes or
when it's launched with a different `meshd` instance.

We have a case like that in gateway: it should be possible to launch
the application without internet access. Since we get our address from
Commissioning, we should be able to do so only when necessary.