models: Support partial configuration updates
As it turns out, UpdateModelConfiguration calls coming from the daemon
might not contain a "complete" model configuration - e.g. when
subscription list has been modified, we won't receive current bindings,
etc.

This resulted in a bug in ModelConfig class and
Model.update_configuration function: when we add bindings, subscription
list got cleared, and when we add subscriptions, bindings get cleared...

This caused app keys to be needlessly added and bound back and forth.